+++ plugins/postcss-text-decoration-shorthand/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Text Decoration Shorthand
 
+### Unreleased (major)
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
 ### 4.0.3
 
 _August 22, 2025_